requirements.txt
qdrant-client>=1.1.0
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=1.3.0
matplotlib>=3.5.0
-e .
